+++ CHANGES.md
@@ -2,9 +2,11 @@
 
 
 ## Unreleased
+- Maintenance: Updated to `packaging.version` v26.0
 
 ## 2026-03-31 v0.3.2
 - Maintenance: Added `py.typed` marker file, signalling typing support
+- packaging.version: Updated to packaging v26.0
 
 ## 2025-02-11 v0.3.1
 - Fixed packaging